pyproject.toml
[project]
name = "cvdanger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "aiobotocore>=3.1.3",
    "aiofiles>=25.1.0",
    "aiogram>=3.24.0",
    "asyncpg>=0.31.0",
    "av>=16.1.0",
    "bentoml>=1.4.35",
    "celery>=5.6.2",
    "fastapi[standard]>=0.128.5",
    "faststream[cli,kafka]>=0.6.6",
    "ffmpeg>=1.4",
    "gitkeep>=0.0.3",
    "groq>=1.0.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "strawberry-graphql[fastapi]>=0.291.3",
    "streamlit>=1.54.0",
    "ultralytics>=8.4.14",
]
